+++ packages/presets/src/theme/recipes/input.ts
@@ -20,8 +20,8 @@ export const input = defineRecipe({
       cursor: 'not-allowed',
     },
     _focus: {
-      borderColor: 'border.outline',
-      boxShadow: 'outline',
+      borderColor: 'border.accent',
+      boxShadow: 'accent',
     },
   },
   defaultVariants: {